rag_project/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
groq
python-dotenv
python-jose passlib
pypdf
langchain
langchain-text-splitters
sentence-transformers
python-multipart